refactor(test): split monolithic conftest into focused modules
Split 1,004-line conftest.py into focused modules for better organization:
- conftest.py (96 lines): shared config, constants, pytest_plugins orchestration
- conftest_numpy.py (198 lines): NumPy array fixtures (.npy loaders), Hargreaves fixtures
- conftest_xarray.py (624 lines): xarray DataArray fixtures for climate index tests
- conftest_benchmark.py (102 lines): benchmark-specific fixtures

Improvements:
- Migrate all os.path usage to pathlib.Path (19 occurrences)
- Add return type annotations to all fixtures
- Fix rain_mm_366 scoping: module -> function (matches rain_mm, rain_mm_365)
- Use pytest_plugins list for fixture discovery (preserves autodiscovery)
- Reduce main conftest.py by 90% (965 lines removed)

All 727 tests collect successfully with new structure.
